# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: ONCF-2255-FO ExpressJS API
#during dev, should point to your local machine
#host: pprdweb-2255:8443
#host: localhost:8443
host: oncf-2255-api-dev.azurewebsites.net
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: >
      For accessing the API a valid JWT token must be passed in all the queries
      in

      the 'Authorization' header.



      A valid JWT token is generated by the API and retourned as answer of a
      call

      to the route /login giving a valid user & password.



      The following syntax must be used in the 'Authorization' header :

          Bearer xxxxxx.yyyyyyy.zzzzzz
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
...
  /Claim/CreateClaim:
    x-swagger-router-controller: "claim"
    post:
      tags:
      - "Claim"
      operationId: createClaim
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: "body"
        name: body
        required: true
        schema:
          $ref: "#/definitions/createClaimModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/GetTicketAttachment:
    x-swagger-router-controller: "claim"
    get:
      tags:
      - "Claim"
      operationId: getTicketAttachment
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - name: "code"
        in: "query"
        required: true
        type: number
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/GetClaimAttachment:
    x-swagger-router-controller: "claim"
    get:
      tags:
      - "Claim"
      operationId: getClaimAttachment
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - name: "code"
        in: "query"
        required: true
        type: number
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/UpdateTicketAttachment:
    x-swagger-router-controller: "claim"
    patch:
      tags:
      - "Claim"
      operationId: updateTicketAttachment
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: "body"
        name: body
        required: true
        schema:
          $ref: "#/definitions/updateTicketAttachmentModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/UpdateClaimAttachment:
    x-swagger-router-controller: "claim"
    patch:
      tags:
      - "Claim"
      operationId: updateClaimAttachment
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: "body"
        name: body
        required: true
        schema:
          $ref: "#/definitions/updateClaimAttachmentModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/UpdateClaim:
    x-swagger-router-controller: "claim"
    post:
      tags:
      - "Claim"
      operationId: updateClaim
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - in: "body"
        name: body
        required: true
        schema:
          $ref: "#/definitions/updateClaimModel"
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/MyClaims:
    x-swagger-router-controller: "claim"
    get:
      tags:
      - "Claim"
      operationId: myClaims
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /Claim/{code}:
    x-swagger-router-controller: "claim"
    get:
      tags:
      - "Claim"
      operationId: getClaimByCode
      security:
        - Bearer: []
      x-security-scopes:  
        - customer
      parameters:
      - name: "code"
        in: "path"
        required: true
        type: number
      - name: Accept-Language
        in: header
        required: true
        type: string
        enum:
          - "fr-FR"
          - "en-US"
          - "ar-MA"
      responses:
        "200":
          description: Success
          schema:
            # a pointer to a definition
            $ref: "#/definitions/ONCFResponse"
        # responses may fall through to errors
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
...
definitions:
  ONCFResponse:
    required:
      - data
      - error
      - isSuccess
    properties:
      data:
        type: object
      error:
        type: object
      isSuccess:
        type: boolean
  ErrorResponse:
    required:
      - data
      - error
      - isSuccess
    properties:
      data:
        type: object
      error:
        type: object
      isSuccess:
        type: boolean
...
  createClaimModel:
    required:
    - "arrivalStationCode"
    - "categoryCode"
    - "departureStationCode"
    - "subCategoryCode"
    - "ticketAttachment"
    - "trainNumber"
    - "travelDate"
    - "userCode"
    properties:
      departureStationCode:
        type: string
      arrivalStationCode:
        type: string
      trainNumber:
        type: string
      travelDate:
        type: "string"
      categoryCode:
        type: string
      subCategoryCode:
        type: string
      subSubCategoryCode:
        type: string
      claimSubject:
        type: "string"
      claimDetails:
        type: "string"
      ticketAttachment:
        type: "string"
      claimAttachment:
        type: "string"
      isONCFUser:
        type: "boolean"
      userEmail:
        type: string
      userCode:
        type: string
  updateClaimModel:
    required:
    - "code"
    properties:
      code:
        type: number
      claimDetails:
        type: "string"
  updateTicketAttachmentModel:
     properties:
      code:
        type: number
      attachment:
        type: "string"
  updateClaimAttachmentModel:
     properties:
      code:
        type: number
      attachment:
        type: string
...